+++ srcs/objects/string.c
@@ -1,5 +1,6 @@
 #include "bsq.h"
 
+// Creates a new `t_str` structure with the given string and length
 t_str	*str_new(char *str, size_t len)
 {
 	t_str	*res;
@@ -12,6 +13,7 @@ t_str	*str_new(char *str, size_t len)
 	return (res);
 }
 
+// Frees a `t_str` structure and sets its pointer to `NULL`
 t_str	*str_free(t_str **str)
 {
 	if (!str || !*str)